# src/structs/http.type.go
package types

type Product_POST_Body struct {
}
type Product_PATCH_Body struct {
}

type Product_GET_Query struct {
}

type Product_DELETE_Query struct {
}

type Product_DELETE_Body struct {
}

type Response struct {
}